Reorganize `theme_importer`
