+++ crates/theme_importer/src/vscode/syntax.rs
